feat: add findEmails utility function
