+++ .prettierignore
@@ -1,2 +1,3 @@
+/dist
 .eslintcache
 pnpm-lock.yaml
